Update filesystem based functions to pass all tests

+++ backup_app.py
@@ -20,7 +20,15 @@ import send2trash
 SHALLOW = True
 
 def copy_files_from_a_to_b(dira, dirb, files, overwrite=False):
-    """Copy files from source directory to backup directory."""
+    """Copy files from source directory to backup directory.
+
+    Arguments:
+    dira -- source folder path
+    dirb -- backup folder path
+    files -- a list of relative filepaths for files to delete
+
+    Return list of files that were not copied.
+    """
     failed = []
     for filename in files:
         old_path = os.path.join(dira, filename)
@@ -50,7 +58,7 @@ def copy_files_from_a_to_b(dira, dirb, files, overwrite=False):
                 logging.info('Copied!')
         else:
             logging.info('%s --> %s', old_path, new_path)
-            logging.error("Error! No file found!")
+            logging.warning("Warning! No file found!")
             failed.append(filename)
     logging.info('done copying files')
     return failed
@@ -65,9 +73,12 @@ def delete_files_from_b(dirb, files):
     failed = []
     for filepath in files:
         path = os.path.join(dirb, filepath)
-        logging.info("deleting %s", path)
-        logging.info(path := os.path.normpath(path))
-        send2trash.send2trash(path)
+        if os.path.exists(path):
+            logging.info("deleting %s", path)
+            logging.info(path := os.path.normpath(path))
+            send2trash.send2trash(path)
+        else:
+            failed.append(filepath)
     logging.info("Done deleting")
     return failed
 
@@ -75,7 +86,7 @@ def move_files_in_b(file_sets):
     """Move files from one location to another.
 
     Arguments:
-    file_sets -- a list of tuples of filepaths (source_path, destination_path)
+    file_sets -- a list of tuples of abs filepaths (source_path, destination_path)
     """
     failed = []
     for file_set in file_sets:
@@ -84,15 +95,19 @@ def move_files_in_b(file_sets):
 ##        dest = os.path.sep.join(dest[0:-1])
         dest = os.path.sep.join(dest)
 ##        move_file_from_a_to_b(src,dest,fn_src)
-        logging.info('moving FROM:\n%s\nTO:\n%s\n\n\n', src, dest)
-        shutil.move(src, dest)
+        if os.path.exists(dest):
+            failed.append(file_set)
+        else:
+            logging.info('moving FROM:\n%s\nTO:\n%s\n\n\n', src, dest)
+            shutil.move(src, dest)
     logging.info('done moving')
     return failed
 
 def update_files_a_to_b(dira, dirb, files):
     """Copy files from source directory to backup directory."""
-    copy_files_from_a_to_b(dira, dirb, files, overwrite=True)
+    failed = copy_files_from_a_to_b(dira, dirb, files, overwrite=True)
     logging.info("Done updating")
+    return failed
 
 class Report:
     """Stores a directory scan report."""
